+++ app/routes/components/Select/changelog.codegen.ts
@@ -13,4 +13,6 @@ export const changelog = ''
   + '- 2025-10-27: Update styles when Select is opened or focused. Make its\n'
   + '  `min-width` be `8rem` now (rather than `16rem`).\n'
   + '- 2025-10-28: Refactor code to make code cleaner. Add a small gap in trigger.\n'
+  + '- 2025-10-30: Add a `ResizeObserver` to the `Select.Trigger` to update the\n'
+  + "  `Select.List`'s width when the trigger size changes.\n"
   ;