+++ src/components/Select/CHANGELOG.md
@@ -9,3 +9,5 @@
 - 2025-10-27: Update styles when Select is opened or focused. Make its
   `min-width` be `8rem` now (rather than `16rem`).
 - 2025-10-28: Refactor code to make code cleaner. Add a small gap in trigger.
+- 2025-10-30: Add a `ResizeObserver` to the `Select.Trigger` to update the
+  `Select.List`'s width when the trigger size changes.